requirements.txt
Django==3.1.6
python-dotenv==0.15.0
psycopg2-binary==2.9.9
django-dirtyfields==1.7.0
django-safedelete==1.0.0
djangoql==0.17.1
django-searchable-encrypted-fields==0.1.9
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
django-health-check==3.16.3
dj-rest-auth==2.1.3
requests==2.28.2
requests-oauthlib==1.3.0
django-allauth==0.44.0
django-rest-passwordreset==1.1.0
PyJWT==1.7.1
django-cors-headers==3.7.0
django-filter==2.4.0
stripe==5.2.0
python-dateutil==2.8.2
django-admin-inline-paginator==0.3.0
django-s3-storage==0.13.4
django-storages==1.12.3
gunicorn==21.2.0
django-ses==1.0.3
celery==5.1.2
gevent==24.2.1
redis==4.3.1
flower==1.1.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
zenpy==2.0.24
sentry-sdk==1.17.0
pydub==0.25.1
openai==1.12.0
django-ipware==6.0.4
google-ads==23.1.0